refactor: centralize hashing logic into one file and better admission comments
